moving spec and test to explicit runs to allow unit test runs
