+++ Taskfile.yml
@@ -60,8 +60,9 @@ tasks:
   test:
     deps: [clean, agent]
     cmds:
-      - sudo go run github.com/onsi/ginkgo/v2/ginkgo -r --randomize-all --randomize-suites --vv --trace --keep-going ./spec #--cover --coverprofile=.coverage-report.out
-      - go test -tags netgo -ldflags '-extldflags "-static"'-v -race ./test
+      - go test -v ./...
+      - sudo go run github.com/onsi/ginkgo/v2/ginkgo -r --randomize-all --randomize-suites --vv --trace --keep-going ./_spec #--cover --coverprofile=.coverage-report.out
+      - go test -tags netgo -ldflags '-extldflags "-static"'-v -race ./_test
   
   otel_start:
     dir: _scripts/otel
