+++ _spec/node_windows.go
@@ -1,0 +1,7 @@
+package spec
+
+import "fmt"
+
+func init() {
+	fmt.Println("no test to run")
+}
