--- spec/node_windows.go
@@ -1,7 +1,0 @@
-package spec
-
-import "fmt"
-
-func init() {
-	fmt.Println("no test to run")
-}
